# Default implementation for walk