--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,2 @@
+/dist
+/.stack-work