--- a/pandoc-types.cabal
+++ b/pandoc-types.cabal
@@ -28,7 +28,7 @@ License-file:        LICENSE
 Author:              John MacFarlane
 Maintainer:          jgm@berkeley.edu
 Bug-Reports:         https://github.com/jgm/pandoc-types/issues
-Copyright:           (c) 2006-2015 John MacFarlane
+Copyright:           (c) 2006-2017 John MacFarlane
 Category:            Text
 Build-type:          Simple
 Cabal-version:       >=1.8
@@ -52,6 +52,7 @@ Library
                      ghc-prim >= 0.2,
                      bytestring >= 0.9 && < 0.11,
                      aeson >= 0.6.2 && < 1.3,
+                     transformers >= 0.2 && < 0.6,
                      QuickCheck >= 2
   if impl(ghc < 7.10)
     Build-depends:   deepseq-generics >= 0.1 && < 0.2