The implementations for `walk` and `walkM` are very similar, so a default method is provided which implements the former in terms of the latter. This change should not affect performance, as the `Identity` functor, which is used in the default definition, is a newtype that should be eliminated at compile time.

GHC 7.8 (i.e., base <= 4.7) does not provide the `Data.Functor.Identity` type, so we use the one included in the transformers library. The transformers library was already an indirect dependency due to Aeson, so while requiring it is ugly, it does not cause additional code to be included.