@jgm 2017-09-12
This is a nice simplification.  Are you sure it doesn't have performance ramifications? It would be nice to have some kind of benchmark.