@jgm 2017-09-12
10% isn't too bad, so we might think about whether the
simplification is worth it.

+++ Albert Krewinkel [Sep 12 17 22:26 ]:
>   I did a small benchmark using criterion and found a ~10% performance
>   degradation. I don't understand where this comes from. Lesson learned:
>   measuring trumps reasoning.
>
>  Benchmarking code
>
>import Criterion.Main (bench, defaultMain, nf)
>import Data.Aeson (eitherDecode)
>import Text.Pandoc.Definition (Pandoc, Inline (Str))
>import Text.Pandoc.Walk (walk)
>
>import qualified Data.ByteString.Lazy as BSL
>
>main :: IO ()
>main = do
>  pd <- parseManual
>  defaultMain [ bench "walk" $ nf (walk prependZeroWidthSpace) pd]
>
>prependZeroWidthSpace :: Inline -> Inline
>prependZeroWidthSpace (Str s) = Str ('\8203' : s)
>prependZeroWidthSpace x = x
>
>parseManual :: IO Pandoc
>parseManual = do
>  content <- eitherDecode <$> BSL.readFile "manual.json"
>  case content of
>    Right pd -> return pd
>    _ -> error "couldn't read manual"
>
>  Results
>
>    master
>
>benchmarking walk
>time                 2.647 ms   (2.620 ms .. 2.678 ms)
>                     0.999 R²   (0.999 R² .. 1.000 R²)
>mean                 2.658 ms   (2.643 ms .. 2.676 ms)
>std dev              52.16 μs   (36.70 μs .. 78.66 μs)
>
>    This PR
>
>benchmarking walk
>time                 2.931 ms   (2.907 ms .. 2.962 ms)
>                     0.999 R²   (0.999 R² .. 1.000 R²)
>mean                 2.958 ms   (2.939 ms .. 2.988 ms)
>std dev              74.13 μs   (52.64 μs .. 113.0 μs)
>variance introduced by outliers: 11% (moderately inflated)
>
>   —
>   You are receiving this because you commented.
>   Reply to this email directly, [1]view it on GitHub, or [2]mute the
>   thread.
>
>References
>
>   1. https://github.com/jgm/pandoc-types/pull/30#issuecomment-329001800
>   2. https://github.com/notifications/unsubscribe-auth/AAAL5MS4mB-UlpGL-knZzaIwPPuXIO7Wks5shwT3gaJpZM4PSy5Z