@jgm 2017-09-13
I wonder if just adding a SPECIALIZE pragma (for the Identity instance) would close the performance gap?