@tarleb 2017-09-13
Adding `{-# SPECIALIZE Identity :: Inline -> Identity Inline #-}` gave me a syntax error, so I manually added `walk f = runIdentity . walkM (return . f) {-# INLINE walk #-}` to all instance definitions, but that didn't help. 🤔