@tarleb 2017-09-14
As was pointed out to me by @leftaroundabout, enabling GHC optimizations *mostly* fixes this: my benchmarks now show an increase in runtime of only ~4%.