@jgm 2017-09-14
OK, this seems worthwhile, then.